+++ klause/src/commonMain/kotlin/com/eignex/klause/localsearch/DefinitionalSweep.kt
@@ -231,12 +231,12 @@ class DefinitionalSweep internal constructor(
             if (f !is ReifiedLinear) continue
             if (frozenBool(f.auxBoolVar)) continue
             var sum = 0L
-            for (k in f.vars.indices) sum += f.coeffs[k].toLong() * assignment.intValue(f.vars[k])
+            for (k in f.vars.indices) sum += f.coeffs[k] * assignment.intValue(f.vars[k])
             val holds = when (f.op) {
                 LinearOp.LE -> sum <= f.bound
                 LinearOp.GE -> sum >= f.bound
-                LinearOp.EQ -> sum == f.bound.toLong()
-                LinearOp.NE -> sum != f.bound.toLong()
+                LinearOp.EQ -> sum == f.bound
+                LinearOp.NE -> sum != f.bound
             }
             assignment.setBool(f.auxBoolVar, holds)
         }
